# tests/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "CatalogAssetNode": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "host",
              "system",
              "network",
              "service",
              "credential_reference",
              "runbook",
              "decision",
              "project"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "visibility": {
            "const": "detail",
            "default": "detail",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "id",
          "kind",
          "label",
          "status"
        ],
        "title": "CatalogAssetNode",
        "type": "object"
      },
      "CatalogAssetStubNode": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "host",
              "system",
              "network",
              "service",
              "credential_reference",
              "runbook",
              "decision",
              "project"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "visibility": {
            "const": "stub",
            "default": "stub",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "id",
          "kind",
          "label"
        ],
        "title": "CatalogAssetStubNode",
        "type": "object"
      },
...
      "CatalogObjectOut": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/CatalogRecordDiagnostic"
            },
            "title": "Diagnostics",
            "type": "array"
          },
          "health": {
            "anyOf": [
              {
                "enum": [
                  "unknown",
                  "healthy",
                  "degraded",
                  "down",
                  "maintenance"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health"
          },
          "id": {
            "pattern": "^[a-z0-9][a-z0-9_-]*[a-z0-9]$|^[a-z0-9]$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "host",
              "system",
              "network",
              "service",
              "credential_reference",
              "runbook",
              "decision",
              "project"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_changed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Changed"
          },
          "lifecycle": {
            "anyOf": [
              {
                "enum": [
                  "planned",
                  "active",
                  "retired"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle"
          },
          "parent_path": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CatalogAssetNode"
                },
                {
                  "$ref": "#/components/schemas/CatalogAssetStubNode"
                }
              ]
            },
            "title": "Parent Path",
            "type": "array"
          },
          "placement_state": {
            "anyOf": [
              {
                "enum": [
                  "root",
                  "assigned",
                  "unassigned",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement State"
          },
          "provenance": {
            "$ref": "#/components/schemas/CatalogProvenanceOut"
          },
          "record_state": {
            "default": "valid",
            "enum": [
              "valid",
              "corrupt"
            ],
            "title": "Record State",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "inactive",
              "deleted"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "visibility": {
            "const": "detail",
            "default": "detail",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "label",
          "provenance",
          "revision"
        ],
        "title": "CatalogObjectOut",
        "type": "object"
      },
      "CatalogObjectStubOut": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "host",
              "system",
              "network",
              "service",
              "credential_reference",
              "runbook",
              "decision",
              "project"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "parent_path": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CatalogAssetNode"
                },
                {
                  "$ref": "#/components/schemas/CatalogAssetStubNode"
                }
              ]
            },
            "title": "Parent Path",
            "type": "array"
          },
          "placement_state": {
            "anyOf": [
              {
                "enum": [
                  "root",
                  "assigned",
                  "unassigned",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement State"
          },
          "visibility": {
            "const": "stub",
            "default": "stub",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "label"
        ],
        "title": "CatalogObjectStubOut",
        "type": "object"
      },
...
      "CatalogRecordDiagnostic": {
        "properties": {
          "code": {
            "const": "corrupt_record",
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "object_id",
          "message"
        ],
        "title": "CatalogRecordDiagnostic",
        "type": "object"
      },
...
      "Permission": {
        "enum": [
          "discover",
          "read",
          "write",
          "create_child",
          "manage_access",
          "delete"
        ],
        "title": "Permission",
        "type": "string"
      },
...
      "V1ObjectPageOut": {
        "properties": {
          "direction": {
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Direction",
            "type": "string"
          },
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgentCatalogObjectSummary"
                },
                {
                  "$ref": "#/components/schemas/AgentCatalogObjectStub"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "sort": {
            "enum": [
              "id",
              "label",
              "kind",
              "updated_at"
            ],
            "title": "Sort",
            "type": "string"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "required": [
          "items",
          "sort",
          "direction"
        ],
        "title": "V1ObjectPageOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Blockwart",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/objects": {
      "get": {
        "operationId": "get_objects_api_objects_get",
        "parameters": [
          {
            "description": "Exact asset lifecycle",
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planned",
                    "active",
                    "retired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact asset lifecycle",
              "title": "Lifecycle"
            }
          },
          {
            "description": "Exact asset health",
            "in": "query",
            "name": "health",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "healthy",
                    "degraded",
                    "down",
                    "maintenance"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact asset health",
              "title": "Health"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/CatalogObjectOut"
                      },
                      {
                        "$ref": "#/components/schemas/CatalogObjectStubOut"
                      }
                    ]
                  },
                  "title": "Response Get Objects Api Objects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Objects",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/objects/{object_id}": {
      "get": {
        "operationId": "get_object_by_id_api_objects__object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CatalogObjectOut"
                    },
                    {
                      "$ref": "#/components/schemas/CatalogObjectStubOut"
                    }
                  ],
                  "title": "Response Get Object By Id Api Objects  Object Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Object By Id",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/objects": {
      "get": {
        "operationId": "list_v1_objects_api_v1_objects_get",
        "parameters": [
          {
            "description": "Case-insensitive search over id, label, summary, and catalog data",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search over id, label, summary, and catalog data",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "host",
                    "system",
                    "network",
                    "service",
                    "credential_reference",
                    "runbook",
                    "decision",
                    "project"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "description": "Exact typed ancestor reference",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 192,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact typed ancestor reference",
              "title": "Parent"
            }
          },
          {
            "description": "Exact resolved IP address",
            "in": "query",
            "name": "ip",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact resolved IP address",
              "title": "Ip"
            }
          },
          {
            "in": "query",
            "name": "port",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 65535,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Port"
            }
          },
          {
            "description": "Exact endpoint capability",
            "in": "query",
            "name": "endpoint_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact endpoint capability",
              "title": "Endpoint Type"
            }
          },
          {
            "description": "Exact application protocol",
            "in": "query",
            "name": "protocol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact application protocol",
              "title": "Protocol"
            }
          },
          {
            "description": "Exact endpoint exposure",
            "in": "query",
            "name": "exposure",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact endpoint exposure",
              "title": "Exposure"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "inactive",
                    "deleted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planned",
                    "active",
                    "retired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "health",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "healthy",
                    "degraded",
                    "down",
                    "maintenance"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Health"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "manual",
                    "import",
                    "discovery"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "description": "Exact computed freshness state",
            "in": "query",
            "name": "stale",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact computed freshness state",
              "title": "Stale"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor returned by the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor returned by the previous page",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "id",
              "enum": [
                "id",
                "label",
                "kind",
                "updated_at"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "description": "Compute the total matching result count",
            "in": "query",
            "name": "include_total",
            "required": false,
            "schema": {
              "default": false,
              "description": "Compute the total matching result count",
              "title": "Include Total",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ObjectPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List V1 Objects",
        "tags": [
          "api-v1-readonly"
        ]
      }
    },
...
  }
}